requirements-dev.txt
-r requirements.txt
pytest
pre-commit
ruff
pyright
requests-mock
